requirements.txt
Flask
Flask_SQLAlchemy
Flask_Login
Flask_Bcrypt
Flask_WTF
email_validator
pycryptodome
python-dotenv

